Move Viewer component into new file
